+++ src/main/java/es/in2/vcverifier/config/properties/CryptoProperties.java
@@ -1,0 +1,8 @@
+package es.in2.vcverifier.config.properties;
+
+import jakarta.validation.constraints.NotNull;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "crypto")
+public record CryptoProperties(@NotNull String privateKey) {
+}