+++ src/main/java/es/in2/vcverifier/exception/ClientLoadingException.java
@@ -1,0 +1,8 @@
+package es.in2.vcverifier.exception;
+
+public class ClientLoadingException extends RuntimeException{
+
+    public ClientLoadingException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}