+++ src/main/java/es/in2/vcverifier/exception/UnsupportedGrantTypeException.java
@@ -1,0 +1,9 @@
+package es.in2.vcverifier.exception;
+
+public class UnsupportedGrantTypeException extends RuntimeException {
+
+    public UnsupportedGrantTypeException(String message) {
+        super(message);
+    }
+
+}