+++ src/main/java/es/in2/vcverifier/model/credentials/machine/PowerLCMachine.java
@@ -1,0 +1,12 @@
+package es.in2.vcverifier.model.credentials.machine;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+
+@Builder
+public record PowerLCMachine(
+        @JsonProperty("id") String id,
+        @JsonProperty("domain") String domain,
+        @JsonProperty("function") String function,
+        @JsonProperty("action") String action
+) {}