+++ CHANGELOG.md
@@ -4,5 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Pre-release](https://github.com/in2workspace/in2-vc-verifier/releases/tag/v0.1.0): v0.1.0
-- Project scafoldings and setup
+## [v1.0.3](https://github.com/in2workspace/in2-verifier-api/releases/tag/v1.0.3)
+### Added
+- Support for OpenID Connect.
+  - Support Authentication using the Authorization Code Flow.
+  - Support Claims with Requesting Claims using Scope Values (openid learcredential)
+  - Support Passing Request Parameters as JWTs (Passing a Request Object by Reference).
+  - Support Client Authentication method with Private Key JWT.
+  - Support for P-256 ECDSA keys for Signing.
+- Support for OpenID for Verifiable Presentations (OID4VP).
+  - Implement VP Proof of Possession verification.
+  - Implement Issuers, Participants and Services verification against the DOME Trust Framework.
+  - Implement VC verification against the DOME Revoked Credentials List.
+- Implement DOME Human-To-Machine (H2M) authentication.
+  - Implement Login page with QR code.
+- Implement DOME Machine-To-Machine (M2M) authentication.
+- Integrate with the DOME Trust Framework.
+
+### Fixed
+- Fix the issue with Login page not showing Wallet URL.
+- Fix the issue with Login page not valid Registration URL.
+- Fix the issue with Login page not redirecting to the Relying Party after expiration of the QR code.